split and placeed generate and parse into functions, also simplified the api calls
